--- docs/source/reference/grid/structured_quad.rst
@@ -1,78 +1,0 @@
-landlab.grid.structured\_quad package
-=====================================
-
-Submodules
-----------
-
-landlab.grid.structured\_quad.c\_faces module
----------------------------------------------
-
-.. automodule:: landlab.grid.structured_quad.c_faces
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-landlab.grid.structured\_quad.cells module
-------------------------------------------
-
-.. automodule:: landlab.grid.structured_quad.cells
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-landlab.grid.structured\_quad.cfuncs module
--------------------------------------------
-
-.. automodule:: landlab.grid.structured_quad.cfuncs
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-landlab.grid.structured\_quad.faces module
-------------------------------------------
-
-.. automodule:: landlab.grid.structured_quad.faces
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-landlab.grid.structured\_quad.links module
-------------------------------------------
-
-.. automodule:: landlab.grid.structured_quad.links
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-landlab.grid.structured\_quad.nodes module
-------------------------------------------
-
-.. automodule:: landlab.grid.structured_quad.nodes
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-landlab.grid.structured\_quad.rectilinear module
-------------------------------------------------
-
-.. automodule:: landlab.grid.structured_quad.rectilinear
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-landlab.grid.structured\_quad.structured module
------------------------------------------------
-
-.. automodule:: landlab.grid.structured_quad.structured
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: landlab.grid.structured_quad
-   :members:
-   :undoc-members:
-   :show-inheritance:
